+++ backend/ops_api/ops/models/cans.py
@@ -1,6 +1,5 @@
-from typing import Any
-from ops.portfolio.models import Portfolio
-from ops.portfolio.models import portfolio_cans
+from ops.models.portfolios import Portfolio
+from ops.models.portfolios import portfolio_cans
 from ops.utils import BaseModel
 from sqlalchemy import (
     Column,
